requirements.txt
# Core dependencies
requests>=2.31.0
tqdm>=4.66.0
dataclasses>=0.6; python_version < '3.7'

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
google-genai>=1.0.0
together>=1.0.0

# Math and utilities
sympy>=1.12

# Evaluation framework
# Note: Requires lcb_runner package - install separately
# pip install git+https://github.com/LiveCodeBench/LiveCodeBench.git
